requirements-dev.txt
asgiref==3.2.10
Django==3.1.1
ffmpeg==1.4
pytube3==9.6.4
pytz==2020.1
sqlparse==0.3.1
typing-extensions==3.7.4.3
